+++ CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
 **ENHANCEMENTS**
 - Allow configuration of static and dynamic node priorities in compute resources via the ParallelCluster configuration YAML file.
 - Add support for Ubuntu 22.
+- Allow memory-based scheduling when multiple instance types are specified for a Slurm Compute Resource.
 
 **CHANGES**